% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CalCHI.R
\name{CalCHI}
\alias{CalCHI}
\title{Calinski-Harabasz Index Calculation}
\usage{
CalCHI(
  hclust_result,
  dist_matrix = NULL,
  max_clusters = round(1 + 3.3 * log(length(hclust_result$order), 10))
)
}
\arguments{
\item{hclust_result}{A hierarchical clustering object (result of hclust function).}

\item{dist_matrix}{An optional distance matrix. If not provided, the cophenetic matrix is used. The matrix should be of type "euclidean".}

\item{max_clusters}{Integer. The maximum number of clusters to evaluate (default: round(1 + 3.3 * log(length(hclust_result$order), 10))).}
}
\value{
A vector containing the Calinski-Harabasz index values for each number of clusters.
}
\description{
This function calculates the Calinski-Harabasz index for evaluating the quality of clustering solutions. It is used to determine the optimal number of clusters in a hierarchical clustering scenario.
}
\details{
The function calculates the Calinski-Harabasz index for different numbers of clusters to help identify the optimal number of clusters. The index measures the ratio of between-cluster dispersion to within-cluster dispersion.
}
\examples{
# Example usage:
data <- mtcars
dist_matrix <- dist(data)
hclust_result <- hclust(dist_matrix)
calinski_values <- CalCHI(hclust_result, dist_matrix)

}
\author{
Zaoqu Liu; Email: liuzaoqu@163.com
}
