% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Select.Features.R
\name{CV.df}
\alias{CV.df}
\title{Calculate Coefficient of Variation for a Data Frame}
\usage{
CV.df(df)
}
\arguments{
\item{df}{A data frame with row samples and column features.}
}
\value{
A sorted vector of CV values for each column in the data frame, in decreasing order.
}
\description{
This function calculates the coefficient of variation (CV) for each column in a data frame.
}
\author{
Zaoqu Liu; E-mail: liuzaoqu@163.com
}
