% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Select.Features.R
\name{CV}
\alias{CV}
\title{Calculate Coefficient of Variation for a Numeric Vector}
\usage{
CV(V)
}
\arguments{
\item{V}{A numeric vector.}
}
\value{
The coefficient of variation (CV) for the input vector.
}
\description{
This function calculates the coefficient of variation (CV) for a given numeric vector.
}
\author{
Zaoqu Liu; E-mail: liuzaoqu@163.com
}
