% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Find.OptClusterFeatures.R
\name{Find.OptClusterFeatures}
\alias{Find.OptClusterFeatures}
\title{Optimal Feature Combination for Multi-Modality Clustering}
\usage{
Find.OptClusterFeatures(
  data_layers,
  feature_subset_sizes,
  try_num_clusters = 2:6,
  n_runs = 5,
  n_fold = 5
)
}
\arguments{
\item{data_layers}{A named list of matrices representing different modalities. Each matrix should have features as rows and samples as columns. A minimum of two datasets is required.}

\item{feature_subset_sizes}{A list of sequences representing possible feature subset sizes for each modality. The names of this list must match the names in data_layers.}

\item{try_num_clusters}{Integer vector. The range of cluster numbers to be tested (default: 2:6).}

\item{n_runs}{Integer. Number of iterations for NMF for each cluster number (default: 5).}

\item{n_fold}{Integer. Number of folds for cross-validation in NMF (default: 5).}
}
\value{
A list containing:
  - optimal_combination: A dataframe with the optimal feature combination and associated clustering score.
  - all_results: A dataframe containing CPI and GAP scores for all tested feature combinations and cluster numbers.
}
\description{
This function selects the optimal combination of features for multi-modality clustering analysis by integrating various modalities (e.g., mutation, CNV, RNA, protein, pathology, radiology) to explore the optimal number of clusters. It uses Nonnegative Matrix Factorization (NMF) for cluster consistency analysis and Multi-block Principal Component Analysis (mbPCA) for assessing cluster separation. The combination of these methods helps identify distinct biological subgroups and evaluate the stability and biological relevance of clustering.
}
\author{
Zaoqu Liu
}
