% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.binary.clusters.R
\name{get.binary.clusters}
\alias{get.binary.clusters}
\title{Get Binary Clusters from Clustering Results}
\usage{
get.binary.clusters(res)
}
\arguments{
\item{res}{A list containing clustering results}
}
\value{
A data frame where each row represents a binary encoding of cluster assignments across different methods.
}
\description{
This function extracts binary cluster assignments from multiple clustering results.
}
\author{
Zaoqu Liu; Email: liuzaoqu@163.com
}
