% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.class.R
\name{get.class}
\alias{get.class}
\title{Get Cluster Assignments}
\usage{
get.class(cc.res, k)
}
\arguments{
\item{cc.res}{Consensus clustering results from ConsensusClusterPlus.}

\item{k}{Integer. The number of clusters to extract.}
}
\value{
A data frame with the following columns:
  - ID: The sample identifier.
  - Cluster: The assigned cluster label, prefixed by 'C'.
}
\description{
Extract cluster assignments from Consensus Clustering results for a specific number of clusters.
}
\examples{
data <- mtcars
cc_res <- RunCC(data)
clu <- get.class(cc_res, 2)
clu

}
\author{
Zaoqu Liu; Email: liuzaoqu@163.com
}
