% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.Jaccard.Distance.R
\name{get.Jaccard.Distance}
\alias{get.Jaccard.Distance}
\title{Jaccard Distance Calculation for Binary Matrix}
\usage{
get.Jaccard.Distance(data, dissimilarity = TRUE)
}
\arguments{
\item{data}{A binary matrix where rows represent features and columns represent samples.}

\item{dissimilarity}{Logical. If TRUE, returns the Jaccard distance; if FALSE, returns the Jaccard similarity (default: TRUE).}
}
\value{
A matrix containing the Jaccard distance or similarity between each pair of columns in the input matrix.
}
\description{
This function calculates the Jaccard distance or similarity for a binary matrix. It is typically used to evaluate the similarity or dissimilarity between columns of a binary matrix.
}
\details{
The function computes the Jaccard distance (or similarity) between each pair of columns in the input binary matrix. The Jaccard distance is calculated as 1 minus the Jaccard similarity.
}
\examples{
data <- matrix(sample(0:1, 1500, replace = TRUE), nrow = 30, ncol = 50)
jaccard_dist <- get.Jaccard.Distance(as.data.frame(data), dissimilarity = TRUE)
jaccard_dist

}
\author{
Zaoqu Liu; Email: liuzaoqu@163.com
}
