% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Select.Features.R
\name{MAD.df}
\alias{MAD.df}
\title{Calculate Median Absolute Deviation for a Data Frame}
\usage{
MAD.df(df)
}
\arguments{
\item{df}{A data frame with row samples and column features.}
}
\value{
A sorted vector of MAD values for each column in the data frame, in decreasing order.
}
\description{
This function calculates the median absolute deviation (MAD) for each column in a data frame.
}
\author{
Zaoqu Liu; E-mail: liuzaoqu@163.com
}
