% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Select.Features.R
\name{Mean.df}
\alias{Mean.df}
\title{Calculate Mean Value for a Data Frame}
\usage{
Mean.df(df)
}
\arguments{
\item{df}{A data frame with row samples and column features.}
}
\value{
A vector of mean values for each column in the data frame.
}
\description{
This function calculates the mean value for each column in a data frame.
}
\author{
Zaoqu Liu; E-mail: liuzaoqu@163.com
}
