% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Select.Features.R
\name{Median.df}
\alias{Median.df}
\title{Calculate Median Value for a Data Frame}
\usage{
Median.df(df)
}
\arguments{
\item{df}{A data frame with row samples and column features.}
}
\value{
A vector of median values for each column in the data frame.
}
\description{
This function calculates the median value for each column in a data frame.
}
\author{
Zaoqu Liu; E-mail: liuzaoqu@163.com
}
