% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Select.Features.R
\name{minmax.df}
\alias{minmax.df}
\title{Minmax Normalization for a Data Frame}
\usage{
minmax.df(data)
}
\arguments{
\item{data}{A data frame with row samples and column features.}
}
\value{
A data frame with normalized values between 0 and 1 for each column.
}
\description{
This function performs min-max normalization on each column in a data frame.
}
\author{
Zaoqu Liu; E-mail: liuzaoqu@163.com
}
