% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Select.Features.R
\name{minmax}
\alias{minmax}
\title{Minmax Normalization for a Numeric Vector}
\usage{
minmax(x)
}
\arguments{
\item{x}{A numeric vector.}
}
\value{
A normalized numeric vector with values between 0 and 1.
}
\description{
This function performs min-max normalization on a numeric vector.
}
\author{
Zaoqu Liu; E-mail: liuzaoqu@163.com
}
