% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RunBCC.R
\name{RunBCC}
\alias{RunBCC}
\title{Run Bayesian Consensus Clustering (BCC) for Multi-Modality Data Integration}
\usage{
RunBCC(data = NULL, N.clust = NULL, max.iterations = 10)
}
\arguments{
\item{data}{A list of matrices where each element represents a different modality (e.g., RNA, protein, methylation). Each matrix should have rows as features and columns as samples.}

\item{N.clust}{Integer. Number of clusters to create from the BCC components (optional but recommended).}

\item{max.iterations}{Integer. Maximum number of iterations for the Bayesian algorithm (default: 10).}
}
\value{
A data frame with the following columns:
  - Sample: The sample identifier.
  - Cluster: The assigned cluster number for each sample.
  - Cluster2: The assigned cluster label, prefixed by 'BCC' to indicate that the clustering was performed using BCC.
}
\description{
This function performs clustering analysis using Bayesian Consensus Clustering (BCC) to integrate multiple omics datasets. BCC is a Bayesian method that helps capture shared patterns across different datasets by finding a consensus clustering solution.
}
\details{
This function uses BCC to integrate multiple data matrices and assign clusters to the samples. BCC uses a Bayesian approach to identify shared patterns across multiple datasets, providing a robust clustering solution.

The function operates as follows:
1. Each matrix in the input list is converted to a matrix to ensure compatibility.
2. BCC is used to identify shared patterns across different modalities.
3. The function returns a data frame containing the cluster assignment for each sample, along with additional information about the clustering process.
}
\examples{
# Example usage:
data1 <- matrix(rnorm(10000), nrow = 100, ncol = 100)
data2 <- matrix(rnorm(10000), nrow = 100, ncol = 100)
colnames(data1) <- colnames(data2) <- paste0("Sample", 1:100)
data_list <- list(data1, data2)

# Run BCC clustering
result <- RunBCC(data = data_list, N.clust = 3)

}
\references{
Lock EF, Dunson DB. Bayesian consensus clustering. Bioinformatics. 2013;29(20):2610-2616. doi:10.1093/bioinformatics/btt425.
}
\author{
Zaoqu Liu; Email: liuzaoqu@163.com
}
