% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RunCIMLR.R
\name{RunCIMLR}
\alias{RunCIMLR}
\title{Run Consensus Iterative Multi-view Learning (CIMLR) for Multi-Modality Data Integration}
\usage{
RunCIMLR(
  data,
  N.clust = NULL,
  num.dimensions = NA,
  tuning.parameter = 10,
  cores.ratio = 1
)
}
\arguments{
\item{data}{A list of matrices where each element represents a different modality (e.g., RNA, protein, methylation). Each matrix should have rows as features and columns as samples.}

\item{N.clust}{Integer. Number of clusters to create (optional but recommended).}

\item{num.dimensions}{Integer. Number of dimensions for CIMLR (default: NA).}

\item{tuning.parameter}{Integer. Tuning parameter for CIMLR (default: 10).}

\item{cores.ratio}{Numeric. Ratio of the number of cores to be used when computing the multi-kernel (default: 1).}
}
\value{
A data frame with the following columns:
  - Sample: The sample identifier.
  - Cluster: The assigned cluster number for each sample.
  - Cluster2: The assigned cluster label, prefixed by 'CIMLR' to indicate that the clustering was performed using CIMLR.
}
\description{
This function performs clustering analysis using Consensus Iterative Multi-view Learning (CIMLR) to integrate multiple omics datasets. CIMLR is useful for discovering shared patterns across multiple datasets and identifying distinct subtypes.
}
\details{
This function uses CIMLR to integrate multiple data matrices and assign clusters to the samples. CIMLR is particularly effective for multi-view learning and discovering common patterns among different data types.

The function operates as follows:
1. CIMLR is performed using the CIMLR package to extract components that summarize the shared variation across different modalities.
2. Feature ranking is applied to identify the most important features across all data.
3. The function returns a data frame containing the cluster assignment for each sample, along with additional information about the clustering process.
}
\examples{
# Example usage:
data1 <- matrix(rnorm(10000), nrow = 100, ncol = 100)
data2 <- matrix(rnorm(10000), nrow = 100, ncol = 100)
colnames(data1) <- colnames(data2) <- paste0("Sample", 1:100)
data_list <- list(data1, data2)

# Run CIMLR clustering
result <- RunCIMLR(data = data_list, N.clust = 3)

}
\references{
Ramazzotti D, Lal A, Wang B, Batzoglou S, Sidow A. Multi-omic tumor data reveal diversity of molecular mechanisms that correlate with survival. Nat Commun. 2018;9(1):4453. doi:10.1038/s41467-018-06921-8.
}
\author{
Zaoqu Liu; Email: liuzaoqu@163.com
}
