% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RunCPCA.R
\name{RunCPCA}
\alias{RunCPCA}
\title{Run Consensus Principal Component Analysis (CPCA) for Multi-Modality Data Integration}
\usage{
RunCPCA(
  data = NULL,
  N.clust = NULL,
  num.components = 2,
  integration.algorithm = "globalScore",
  nonzero.coeff.k = "all",
  center.data = FALSE,
  scale.data = FALSE,
  normalization.option = "uniform",
  max.iterations = 1000,
  return.moa.object = TRUE,
  show.verbose = FALSE,
  svd.solver.method = "fast.svd",
  nonzero.coeff.obs = "all",
  weight.variables = NA,
  weight.observations = NA,
  unit.length.variables = FALSE,
  unit.length.observations = FALSE,
  retain.nonnegative = FALSE,
  clustering.algorithm = "ward.D2"
)
}
\arguments{
\item{data}{A list of matrices where each element represents a different modality (e.g., RNA, protein, methylation). Each matrix should have rows as features and columns as samples.}

\item{N.clust}{Integer. Number of clusters to create from the hierarchical clustering of the CPCA components (optional but recommended).}

\item{num.components}{Integer. Number of components to retain for each modality (default: 2).}

\item{integration.algorithm}{Character. Algorithm to use for CPCA, options include "globalScore" (default), "blockScore", or "blockLoading".}

\item{nonzero.coeff.k}{Numeric or Character. The number (if >= 1) or proportion (if 0 < k < 1) of non-zero coefficients for variable loadings (default: "all").}

\item{center.data}{Logical. Whether to center each block to zero mean (default: FALSE).}

\item{scale.data}{Logical. Whether to scale each block to unit variance (default: FALSE).}

\item{normalization.option}{Character. Normalization option, one of "lambda1", "inertia", or "uniform" (default: "uniform").}

\item{max.iterations}{Integer. Maximum number of iterations (default: 1000).}

\item{return.moa.object}{Logical. Whether to return an object of class `moa-class` (default: TRUE).}

\item{show.verbose}{Logical. Whether to print process information (default: FALSE).}

\item{svd.solver.method}{Character. SVD solver to use, one of "svd", "fast.svd", or "propack" (default: "fast.svd").}

\item{nonzero.coeff.obs}{Numeric or Character. Number or proportion of non-zero coefficients for observation scores (default: "all").}

\item{weight.variables}{Numeric, Vector, or List. Weights for variables (default: NA).}

\item{weight.observations}{Numeric, Vector, or List. Weights for observations (default: NA).}

\item{unit.length.variables}{Logical. Whether the loading vectors for each block should have unit length (default: FALSE).}

\item{unit.length.observations}{Logical. Whether the score vectors for each block should have unit length (default: FALSE).}

\item{retain.nonnegative}{Logical. Whether to retain only non-negative coefficients in loadings and scores (default: FALSE).}

\item{clustering.algorithm}{Character. The clustering algorithm to use for hierarchical clustering (default: "ward.D2").}
}
\value{
A data frame with the following columns:
  - Sample: The sample identifier.
  - Cluster: The assigned cluster number for each sample.
  - Cluster2: The assigned cluster label, prefixed by 'CPCA' to indicate that the clustering was performed using CPCA.
}
\description{
This function performs clustering analysis using Consensus Principal Component Analysis (CPCA) to integrate multiple omics data. CPCA helps capture the shared variation across different data modalities and provides insights into the common structure of the data.
}
\details{
This function uses CPCA to integrate multiple data matrices and then performs hierarchical clustering on the resulting components to identify clusters.
}
\examples{
# Example usage:
data1 <- matrix(rnorm(10000), nrow = 100, ncol = 100)
data2 <- matrix(rnorm(10000), nrow = 100, ncol = 100)
colnames(data1) <- colnames(data2) <- paste0("Sample", 1:100)
data_list <- list(data1, data2)

# Run CPCA clustering
result <- RunCPCA(data = data_list, N.clust = 3)

}
\references{
Meng C, Basunia A, Peters B, Gholami AM, Kuster B, Culhane AC. MOGSA: Integrative Single Sample Gene-set Analysis of Multiple Omics Data. Mol Cell Proteomics. 2019;18(8 suppl 1):S153-S168. doi:10.1074/mcp.TIR118.001251.
}
\author{
Zaoqu Liu; Email: liuzaoqu@163.com
}
