% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RuniClusterBayes.R
\name{RuniClusterBayes}
\alias{RuniClusterBayes}
\title{Run Bayesian iCluster (iClusterBayes) for Multi-Modality Data Integration}
\usage{
RuniClusterBayes(
  data = NULL,
  N.clust = NULL,
  data.type = c("binomial", "gaussian", "gaussian", "gaussian", "gaussian", "gaussian"),
  num.burnin = 20,
  num.draws = 10,
  prior.probabilities = rep(0.5, length(data)),
  proposal.sdev = 0.05,
  thinning.interval = 3
)
}
\arguments{
\item{data}{A list of matrices where each element represents a different modality (e.g., RNA, protein, methylation). Each matrix should have rows as features and columns as samples.}

\item{N.clust}{Integer. Number of clusters to create (optional but recommended).}

\item{data.type}{A character vector specifying the type of data for each modality. Options include "binomial", "gaussian", etc. Default is c("binomial", "gaussian", "gaussian", "gaussian", "gaussian", "gaussian").}

\item{num.burnin}{Integer. Number of burn-in iterations for the Bayesian algorithm (default: 18000).}

\item{num.draws}{Integer. Number of MCMC draws after burn-in (default: 12000).}

\item{prior.probabilities}{Numeric vector. Prior values for the inclusion probabilities for each modality (default: 0.5 for each modality).}

\item{proposal.sdev}{Numeric. Standard deviation of the proposal distribution (default: 0.05).}

\item{thinning.interval}{Integer. Thinning interval for MCMC sampling (default: 3).}
}
\value{
A data frame with the following columns:
  - Sample: The sample identifier.
  - Cluster: The assigned cluster number for each sample.
  - Cluster2: The assigned cluster label, prefixed by 'iClusterBayes' to indicate that the clustering was performed using iClusterBayes.
}
\description{
This function performs clustering analysis using Bayesian iCluster (iClusterBayes) to integrate multiple omics datasets. iClusterBayes is an extension of iCluster that allows for Bayesian inference, which is useful for identifying shared patterns across multiple datasets with different distributions.
}
\details{
This function uses iClusterBayes to integrate multiple data matrices and assign clusters to the samples. iClusterBayes performs Bayesian inference, which is useful for modeling different data distributions across multiple datasets.

The function operates as follows:
1. Each matrix in the input list is transposed so that rows represent samples and columns represent features.
2. iClusterBayes is used to identify shared patterns across the different modalities.
3. The function returns a data frame containing the cluster assignment for each sample, along with additional information about the clustering process.
}
\examples{
# Example usage:
data1 <- matrix(rnorm(10000), nrow = 100, ncol = 100)
data2 <- matrix(rnorm(10000), nrow = 100, ncol = 100)
colnames(data1) <- colnames(data2) <- paste0("Sample", 1:100)
data_list <- list(data1, data2)

# Run iClusterBayes clustering
result <- RuniClusterBayes(
  data = data_list, N.clust = 3,
  data.type = c("gaussian", "gaussian")
)

}
\references{
Mo Q, Shen R, et al. (2013) Pattern discovery and cancer gene identification in integrated cancer genomic data. PNAS, 110 (11) 4245-4250.
Shen R, Olshen AB, et al. (2012) Integrative Subtype Discovery in Glioblastoma Using iCluster, PLOS ONE 7(4):e35236.
Shen R, Olshen AB, et al. (2009) Integrative clustering of multiple genomic data types using a joint latent variable model with application to breast and lung cancer subtype analysis. Bioinformatics, 25(22):2906-12.
}
\author{
Zaoqu Liu; Email: liuzaoqu@163.com
}
