% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RunNEMO.R
\name{RunNEMO}
\alias{RunNEMO}
\title{Run NEMO for Multi-Modality Data Clustering}
\usage{
RunNEMO(data, N.clust = NULL)
}
\arguments{
\item{data}{A list of matrices where each element represents a different modality (e.g., RNA, protein, methylation). Each matrix should have rows as features and columns as samples.}

\item{N.clust}{Integer. Number of clusters to evaluate during subtyping. If NULL, NEMO will determine the optimal number of clusters automatically. Default is NULL.}
}
\value{
A data frame with the following columns:
  - Sample: The sample identifier, taken from the column names of the input data matrices.
  - Cluster: The assigned cluster number for each sample, indicating the subtype classification.
  - Cluster2: The assigned cluster label, prefixed by 'NEMO' to indicate that the clustering was performed using NEMO.
}
\description{
This function performs clustering analysis using the NEMO method, which is suitable for identifying subtypes in multi-omics data. NEMO allows integration of multiple modalities to discover coherent subtypes by partial data integration.
}
\details{
The function operates as follows:
1. Uses the NEMO package to identify subtypes across the provided omics modalities.
2. Returns a data frame containing the cluster assignment for each sample.
}
\examples{
# Example usage:
data1 <- matrix(rnorm(10000), nrow = 100, ncol = 100)
data2 <- matrix(rnorm(10000), nrow = 100, ncol = 100)
colnames(data1) <- colnames(data2) <- paste0("Sample", 1:100)
data_list <- list(data1, data2)

# Run NEMO clustering
result <- RunNEMO(data = data_list, N.clust = 3)
}
\references{
Rappoport N, Shamir R. NEMO: cancer subtyping by integration of partial multi-omic data. Bioinformatics. 2019;35(18):3348-3356. doi:10.1093/bioinformatics/btz058
}
\author{
Zaoqu Liu; Email: liuzaoqu@163.com
}
