% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RunPCA.R
\name{RunPCA}
\alias{RunPCA}
\title{Run PCA on Data}
\usage{
RunPCA(data)
}
\arguments{
\item{data}{A distance matrix or data frame on which to perform PCA.}
}
\value{
The PCA result as produced by the FactoMineR package.
}
\description{
This function performs Principal Component Analysis (PCA) on the given data using the FactoMineR package.
}
\details{
This function uses the FactoMineR package to perform PCA, which can be used to visualize the relationships between samples in reduced dimensional space.
}
\examples{
# Example usage:
data <- matrix(rnorm(100), nrow = 10, ncol = 10)
pca_result <- RunPCA(data)

}
\author{
Zaoqu Liu; Email: liuzaoqu@163.com
}
