% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Select.Features.R
\name{SD.df}
\alias{SD.df}
\title{Calculate Standard Deviation for a Data Frame}
\usage{
SD.df(df)
}
\arguments{
\item{df}{A data frame with row samples and column features.}
}
\value{
A sorted vector of SD values for each column in the data frame, in decreasing order.
}
\description{
This function calculates the standard deviation (SD) for each column in a data frame.
}
\author{
Zaoqu Liu; E-mail: liuzaoqu@163.com
}
