% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Select.Features.R
\name{Select.Features}
\alias{Select.Features}
\title{Select Hypervariable Features}
\usage{
Select.Features(
  data,
  method = "mad",
  top.percent = NULL,
  top.number = 1000,
  custom.features = NULL
)
}
\arguments{
\item{data}{A data frame with row features and column samples.}

\item{method}{Method of calculating variance ("sd", "mad", "cv").}

\item{top.percent}{The top percent of hypervariable features based on variance.}

\item{top.number}{The top number of hypervariable features based on variance.}

\item{custom.features}{A vector of custom features to select.}
}
\value{
A vector of selected hypervariable features.
}
\description{
This function selects hypervariable features from a data frame based on specified variance calculation methods.
}
\examples{
\dontrun{
df <- data.frame(matrix(rnorm(1000), nrow = 100, ncol = 10))
selected_features <- Select.Features(df, method = "mad", top.percent = 10)
}
}
\author{
Zaoqu Liu; E-mail: liuzaoqu@163.com
}
